# frontend/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/auth/code": {
      "post": {
        "summary": "sendCode",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": false
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/register": {
      "post": {
        "summary": "注册新用户。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": false
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/auth/login": {
      "post": {
        "summary": "用户登录。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "string",
            "description": "具体数据内容"
          }
        }
      },
...
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "boolean",
            "description": "具体数据内容"
          }
        }
      },
...
      "EmailCodeRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "",
            "format": "email"
          },
          "scene": {
            "type": "string",
            "description": "",
            "pattern": "REGISTER|LOGIN"
          }
        },
        "required": [
          "email",
          "scene"
        ]
      },
      "UserRegisterDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "",
            "minLength": 6,
            "maxLength": 32
          },
          "nickname": {
            "type": "string",
            "description": ""
          },
          "code": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "email",
          "password",
          "nickname",
          "code"
        ]
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}